test: fix partner logo issue
