+++ src/components/Footer/index.js
@@ -1,51 +1,29 @@
 import React from 'react'
 import { graphql, useStaticQuery } from 'gatsby'
-const _partners = [
-  {
-    image: '/img/partner-logos-vas.svg',
-    url: 'https://web.vas.com',
-  },
-  {
-    image: '/img/partner-logos-lely.svg',
-    url: 'https://lely.com',
-  },
-  {
-    image: '/img/partner-logos-landolakes.svg',
-    url: 'https://www.landolakes.com',
-  },
-  {
-    image: '/img/partner-logos-nestle.svg',
-    url: 'https://www.nestle.com',
-  },
-  {
-    image: '/img/partner-logos-allflex.svg',
-    url: 'http://www.allflexusa.com',
-  },
-]
 
-const Partners = () => {
-  const data = useStaticQuery(
-    graphql`
-      query PartnerData {
-        allMarkdownRemark(filter: { frontmatter: { footer: { partners: { elemMatch: { image: { ne: null } } } } } }) {
-          nodes {
-            frontmatter {
-              footer {
-                partners {
-                  image
-                  link
-                }
-              }
-            }
-          }
-        }
-      }
-    `
-  )
-  return data.allMarkdownRemark.nodes[0].frontmatter.footer
-}
+// const Partners = () => {
+//   const data = useStaticQuery(
+//     graphql`
+//       query PartnerData {
+//         allMarkdownRemark(filter: { frontmatter: { footer: { partners: { elemMatch: { image: { ne: null } } } } } }) {
+//           nodes {
+//             frontmatter {
+//               footer {
+//                 partners {
+//                   image
+//                   link
+//                 }
+//               }
+//             }
+//           }
+//         }
+//       }
+//     `
+//   )
+//   return data.allMarkdownRemark.nodes[0].frontmatter.footer
+// }
 
-const Footer = () => {
+const Footer = ({ Partners }) => {
   const { partners } = Partners()
   return (
     <footer className="footer">
@@ -54,7 +32,7 @@ const Footer = () => {
         <div className="footer-content">
           <div className="footer-partners">
             {partners.map(({ image, link }) => (
-              <div className="partner-logo">
+              <div key={link} className="partner-logo">
                 <a href={link}>
                   <img src={image} alt={link} />
                 </a>
